requirements.txt
ollama
# psycopg2 # For production
psycopg2-binary # For Development
datetime